+++ wp-content/themes/core/components/blocks/stats/Stats_Block_Controller.php
@@ -160,7 +160,7 @@ class Stats_Block_Controller extends Abstract_Controller {
 			Text_Controller::CLASSES => [
 				'c-block__leadin',
 				'b-stats__leadin',
-				'tribe-acf-hide-label',
+				'h6',
 			],
 			Text_Controller::CONTENT => $this->leadin ?? '',
 		] );